refactor: use custom hook for company profile fetching
